Editor: implemented UpgradeGameCommonTask, move upgrade code from Tasks
